requirements.txt
anyio==4.4.0
APScheduler==3.10.4
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-telegram-bot==21.4
pytz==2024.1
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sos==4.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
anyio==4.4.0
APScheduler==3.10.4
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mysql-connector-python==9.2.0
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-telegram-bot==21.4
pytz==2024.1
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sos==4.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
anyio==4.4.0
apparmor==3.0.4
APScheduler==3.10.4
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
LibAppArmor==3.0.4
MarkupSafe==2.0.1
more-itertools==8.10.0
mysql-connector-python==9.2.0
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-telegram-bot==21.4
pytz==2024.1
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sos==4.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
